fix(msvc): cast (void) nos Result descartados de secrets_.store (C4834 /WX)
std::expected é [[nodiscard]] no MSVC (o clang do mac não avisa). Descartar o
retorno de secrets_.store em SettingsController (setter + migração) e
SecretsController::remove quebrava o build no /WX. Também limpa clampDim e o
indexOf do CodeHighlighter p/ evitar narrowing (defensivo).

+++ ui/src/CodeHighlighter.cpp
@@ -90,7 +90,7 @@ protected:
         setCurrentBlockState(0);
         static const QRegularExpression startRe("/\\*");
         static const QRegularExpression endRe("\\*/");
-        int start = (previousBlockState() == 1) ? 0 : text.indexOf(startRe);
+        int start = (previousBlockState() == 1) ? 0 : static_cast<int>(text.indexOf(startRe));
         while (start >= 0) {
             const auto end = endRe.match(text, start);
             int len = 0;
